requirements.txt
boto==2.49.0
boto3==1.9.213
botocore==1.12.213
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-social-share==1.3.2
django-social-widgets==0.4.0
django-storages==1.7.1
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
docutils==0.15.2
dropbox==9.4.0
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
jsmin==2.2.2
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
whitenoise==4.1.3
